[refactor] HomeVC와 HomeView를 분리
