+++ Soulmate/Soulmate/Presentation/HomeScene/ViewController/HomeViewController.swift
@@ -13,8 +13,13 @@ final class HomeViewController: UIViewController {
     
     var cancellables = Set<AnyCancellable>()
 
+    // MARK: - ViewModel
     private var viewModel: HomeViewModel?
     
+    // MARK: - UI
+    private var homeView: HomeView?
+    
+    // MARK: - DIffable DataSource
     enum SectionKind: Int, CaseIterable {
         case main
     }
@@ -22,59 +27,16 @@ final class HomeViewController: UIViewController {
     enum ItemKind: Hashable {
         case main(HomePreviewViewModelWrapper)
     }
+    
+    private var dataSource: UICollectionViewDiffableDataSource<SectionKind, ItemKind>?
 
+    // MARK: Subject
+    
     var refreshButtonTapSubject = PassthroughSubject<Void, Never>()
     var collectionViewSelectSubject = PassthroughSubject<Int, Never>()
     var tokenUpdateSubject = PassthroughSubject<String, Never>()
     
-    // MARK: - UI
-    private lazy var logo: UIImageView = {
-        let imageView = UIImageView()
-        imageView.image = UIImage(named: "logo")
-        imageView.contentMode = .scaleAspectFit
-        self.view.addSubview(imageView)
-        return imageView
-    }()
-    
-    private lazy var numOfHeartButton: UIButton = {
-        let button = UIButton()
-        button.setTitleColor(UIColor.labelDarkGrey, for: .normal)
-        button.setImage(UIImage(named: "heart"), for: .normal)
-        button.titleLabel?.font = UIFont(name: "AppleSDGothicNeo-Bold", size: 15)
-        button.contentHorizontalAlignment = .right
-        button.titleEdgeInsets = UIEdgeInsets(top: 0, left: 10, bottom: 0, right: 0)
-        
-        self.view.addSubview(button)
-
-        return button
-    }()
-    
-    private lazy var collectionView: UICollectionView = {
-        let layout = createCompositionalLayout()
-        let collection = UICollectionView(frame: .zero, collectionViewLayout: layout)
-        collection.showsVerticalScrollIndicator = false
-        collection.bounces = true
-        collection.delaysContentTouches = false
-        collection.isPagingEnabled = false
-        collection.backgroundColor = .clear
-        self.view.addSubview(collection)
-        return collection
-    }()
-    
-    private lazy var hiddenLabel: UILabel = {
-        let label = UILabel()
-        label.text = "가까운 거리에 추천 상대가 없어요."
-        label.font = UIFont(name: "AppleSDGothicNeo-Bold", size: 15)
-        label.textColor = .gray
-        label.textAlignment = .center
-        label.layer.cornerRadius = 10
-        label.backgroundColor = .white
-        label.isHidden = true
-        self.view.addSubview(label)
-        return label
-    }()
-    
-    private var dataSource: UICollectionViewDiffableDataSource<SectionKind, ItemKind>?
+    // MARK: Init
     
     override init(nibName nibNameOrNil: String?, bundle nibBundleOrNil: Bundle?) {
         super.init(nibName: nil, bundle: nil)
@@ -89,17 +51,23 @@ final class HomeViewController: UIViewController {
         self.viewModel = viewModel
     }
     
-    // MARK: - 초기화
+    // MARK: LifeCycle
+    
     override func viewDidLoad() {
         super.viewDidLoad()
-        
-        configureView()
-        configureLayout()
-        
-        bind()
+        configure()
         configureDataSource()
-        
+        bind()
+
         updateToken()
+    }
+    
+    override func loadView() {
+        super.loadView()
+        
+        let view = HomeView(frame: self.view.frame)
+        self.view = view
+        self.homeView = view
     }
 
     override func viewWillAppear(_ animated: Bool) {
@@ -120,25 +88,88 @@ final class HomeViewController: UIViewController {
     }
 }
 
-// MARK: - Read UserDefault
+// MARK: - Configure
 private extension HomeViewController {
     
-    func updateToken() {
-        print("homeVC update token")
-        UserDefaults.standard
-            .publisher(for: \.token)
+    func configure() {
+        homeView?.collectionView.delegate = self
+    }
+    
+    func bind() {
+        guard let viewModel = viewModel,
+              let homeView = homeView else { return }
+        
+        let output = viewModel.transform(
+            input: HomeViewModel.Input(
+                viewDidLoad: Just(()).eraseToAnyPublisher(),
+                didTappedRefreshButton: refreshButtonTapSubject.eraseToAnyPublisher(),
+                didSelectedMateCollectionCell: collectionViewSelectSubject.eraseToAnyPublisher(),
+                didTappedHeartButton: homeView.numOfHeartButton.tapPublisher().eraseToAnyPublisher(),
+                tokenUpdateEvent: tokenUpdateSubject.eraseToAnyPublisher()
+            )
+        )
+
+        output.didRefreshedPreviewList
             .compactMap { $0 }
-            .sink { [weak self] token in
-                self?.tokenUpdateSubject.send(token)
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self, homeView] previewViewModelList in
+                if previewViewModelList.count == 0 {
+                    homeView.collectionView.isHidden = true
+                    homeView.hiddenLabel.isHidden = false
+                } else {
+                    homeView.collectionView.isHidden = false
+                    homeView.hiddenLabel.isHidden = true
+                }
+
+                var snapshot = NSDiffableDataSourceSectionSnapshot<ItemKind>()
+                snapshot.append(previewViewModelList.enumerated().map { index, value in
+                    return ItemKind.main(HomePreviewViewModelWrapper(index: index, previewViewModel: value))
+                })
+                self?.dataSource?.apply(snapshot, to: .main) {
+                    homeView.collectionView.isScrollEnabled = true
+                    homeView.collectionView.allowsSelection = true
+                }
+            }
+            .store(in: &cancellables)
+        
+        output.didUpdatedHeartInfo
+            .compactMap { $0 }
+            .receive(on: DispatchQueue.main)
+            .sink { [weak homeView] heartInfo in
+                guard let heart = heartInfo.heart else { return }
+                homeView?.numOfHeartButton.setTitle("\(heart)", for: .normal)
+            }
+            .store(in: &cancellables)
+        
+        output.didStartRefreshing
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] in
+                self?.fakeSnapshot()
+            }
+            .store(in: &cancellables)
+        
+        output.lessHeart
+            .receive(on: DispatchQueue.main)
+            .sink { [weak self] in
+                self?.showPopUp(title: "하트 부족",
+                                message: "하트가 부족합니다. 충전하러 갈까요?",
+                                leftActionTitle: "취소",
+                                rightActionTitle: "충전하기",
+                                rightActionCompletion: {
+                    self?.viewModel?.actions?.showHeartShopFlow?()
+                })
             }
             .store(in: &cancellables)
     }
 }
 
+// MARK: - CollectionView DataSource
 
-// MARK: - DataSource
 private extension HomeViewController {
+
     func configureDataSource() {
+        
+        guard let homeView = homeView else { return }
 
         // MARK: Cell Registration
         
@@ -167,7 +198,7 @@ private extension HomeViewController {
         }
         
         // MARK: DataSource Configuration
-        self.dataSource = UICollectionViewDiffableDataSource<SectionKind, ItemKind>(collectionView: self.collectionView) { (collectionView, indexPath, item) -> UICollectionViewCell? in
+        self.dataSource = UICollectionViewDiffableDataSource<SectionKind, ItemKind>(collectionView: homeView.collectionView) { (collectionView, indexPath, item) -> UICollectionViewCell? in
             switch item {
             case .main(let previewViewModel):
                 return collectionView.dequeueConfiguredReusableCell(using: previewCellRegistration, for: indexPath, item: previewViewModel)
@@ -184,43 +215,14 @@ private extension HomeViewController {
         
         fakeSnapshot()
     }
-    
-    private func createCompositionalLayout() -> UICollectionViewCompositionalLayout {
-        return UICollectionViewCompositionalLayout { sectionNumber, _ -> NSCollectionLayoutSection? in
-            switch sectionNumber {
-            case 0: return self.mainSectionLayout()
-            default: fatalError()
-            }
-        }
-    }
-    
-    private func mainSectionLayout() -> NSCollectionLayoutSection {
-        let itemSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalHeight(1))
-        let item = NSCollectionLayoutItem(layoutSize: itemSize)
-        item.contentInsets = .init(top: 0, leading: 0, bottom: 0, trailing: 0)
-        let groupSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .fractionalWidth(1))
-        let group = NSCollectionLayoutGroup.vertical(layoutSize: groupSize, subitems: [item])
-        group.contentInsets = .init(top: 5, leading: 5, bottom: 5, trailing: 5)
-        
-        let section = NSCollectionLayoutSection(group: group)
-        section.contentInsets = NSDirectionalEdgeInsets(top: 0, leading: 0, bottom: 0, trailing: 0)
-        let footerSize = NSCollectionLayoutSize(widthDimension: .fractionalWidth(1), heightDimension: .absolute(94))
-        let footer = NSCollectionLayoutBoundarySupplementaryItem(
-            layoutSize: footerSize,
-            elementKind: RecommendFooterView.footerKind,
-            alignment: .bottom
-        )
-        footer.contentInsets = .init(top: 0, leading: 5, bottom: 0, trailing: 5)
-        section.boundarySupplementaryItems = [footer]
-
-        return section
-    }
 }
 
 // MARK: CollectionView Skeleton Diffable Snapshot
 private extension HomeViewController {
     
     func fakeSnapshot() {
+        guard let homeView = homeView else { return }
+        
         let estimatedNumberOfRows = Int(ceil(self.view.frame.height / (self.view.frame.width - 20)))
         var snapshot = NSDiffableDataSourceSectionSnapshot<ItemKind>()
         var targetSource = (0..<estimatedNumberOfRows).map {
@@ -228,128 +230,12 @@ private extension HomeViewController {
         }
         snapshot.append(targetSource.map { return .main($0) })
 
-        dataSource?.apply(snapshot, to: .main) { [weak self] in
-            self?.collectionView.isScrollEnabled = false
-            self?.collectionView.allowsSelection = false
+        dataSource?.apply(snapshot, to: .main) { [weak homeView] in
+            homeView?.collectionView.isScrollEnabled = false
+            homeView?.collectionView.allowsSelection = false
         }
-        self.collectionView.scrollToItem(at: IndexPath(row: 0, section: 0), at: .top, animated: true)
+        homeView.collectionView.scrollToItem(at: IndexPath(row: 0, section: 0), at: .top, animated: true)
     }
-}
-
-// MARK: - View Generators
-private extension HomeViewController {
-    func bind() {
-        guard let viewModel = viewModel else { return }
-        
-        let output = viewModel.transform(
-            input: HomeViewModel.Input(
-                viewDidLoad: Just(()).eraseToAnyPublisher(),
-                didTappedRefreshButton: refreshButtonTapSubject.eraseToAnyPublisher(),
-                didSelectedMateCollectionCell: collectionViewSelectSubject.eraseToAnyPublisher(),
-                didTappedHeartButton: numOfHeartButton.tapPublisher().eraseToAnyPublisher(),
-                tokenUpdateEvent: tokenUpdateSubject.eraseToAnyPublisher()
-            )
-        )
-
-        output.didRefreshedPreviewList
-            .compactMap { $0 }
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] previewViewModelList in
-                if previewViewModelList.count == 0 {
-                    self?.collectionView.isHidden = true
-                    self?.hiddenLabel.isHidden = false
-                } else {
-                    self?.collectionView.isHidden = false
-                    self?.hiddenLabel.isHidden = true
-                }
-
-                var snapshot = NSDiffableDataSourceSectionSnapshot<ItemKind>()
-                snapshot.append(previewViewModelList.enumerated().map { index, value in
-                    return ItemKind.main(HomePreviewViewModelWrapper(index: index, previewViewModel: value))
-                })
-                self?.dataSource?.apply(snapshot, to: .main) { [weak self] in
-                    self?.collectionView.isScrollEnabled = true
-                    self?.collectionView.allowsSelection = true
-                }
-            }
-            .store(in: &cancellables)
-        
-        output.didUpdatedHeartInfo
-            .compactMap { $0 }
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] heartInfo in
-                guard let heart = heartInfo.heart else { return }
-                self?.numOfHeartButton.setTitle("\(heart)", for: .normal)
-            }
-            .store(in: &cancellables)
-        
-        output.didStartRefreshing
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] in
-                self?.fakeSnapshot()
-            }
-            .store(in: &cancellables)
-        
-        output.lessHeart
-            .receive(on: DispatchQueue.main)
-            .sink { [weak self] in
-                self?.showPopUp(title: "하트 부족",
-                                message: "하트가 부족합니다. 충전하러 갈까요?",
-                                leftActionTitle: "취소",
-                                rightActionTitle: "충전하기",
-                                rightActionCompletion: {
-                    self?.viewModel?.actions?.showHeartShopFlow?()
-                })
-            }
-            .store(in: &cancellables)
-    }
-    
-    func configureView() {
-        view.backgroundColor = .white
-        collectionView.delegate = self
-    }
-    
-    func configureLayout() {
-        logo.snp.makeConstraints {
-            $0.leading.equalToSuperview().inset(20)
-            $0.top.equalTo(view.snp.top).inset(64)
-            $0.width.equalTo(140)
-        }
-
-        numOfHeartButton.snp.makeConstraints {
-            $0.trailing.equalToSuperview().inset(20)
-            $0.centerY.equalTo(logo.snp.centerY)
-            $0.height.equalTo(28)
-            $0.width.equalTo(100)
-        }
-        
-        collectionView.snp.makeConstraints {
-            $0.top.equalTo(logo.snp.bottom).offset(5)
-            $0.bottom.equalToSuperview()
-            $0.centerX.equalToSuperview()
-            $0.width.equalToSuperview().inset(20)
-        }
-        
-        hiddenLabel.snp.makeConstraints {
-            $0.centerX.centerY.equalToSuperview()
-            $0.height.equalTo(54)
-        }
-    }
-    
-    func setAuthAlertAction() {
-         let authAlertController = UIAlertController(
-            title: "위치 사용 권한이 필요합니다.",
-            message: "위치 권한을 허용해야만 앱을 사용하실 수 있습니다.", preferredStyle: .alert
-         )
-         
-         let getAuthAction = UIAlertAction(title: "설정", style: .default, handler: { (UIAlertAction) in
-             if let appSettings = URL(string: UIApplication.openSettingsURLString) {
-                 UIApplication.shared.open(appSettings,options: [:],completionHandler: nil)
-             }
-         })
-         authAlertController.addAction(getAuthAction)
-         self.present(authAlertController, animated: true, completion: nil)
-     }
 }
 
 // MARK: - CollectionView Delegate
@@ -362,5 +248,22 @@ extension HomeViewController: UICollectionViewDelegate {
         default:
             fatalError("indexPath.section")
         }
+    }
+}
+
+
+// MARK: - Read UserDefault
+
+private extension HomeViewController {
+    
+    func updateToken() {
+        print("homeVC update token")
+        UserDefaults.standard
+            .publisher(for: \.token)
+            .compactMap { $0 }
+            .sink { [weak self] token in
+                self?.tokenUpdateSubject.send(token)
+            }
+            .store(in: &cancellables)
     }
 }
